Crate hf_hub

This crates aims to emulate and be compatible with the huggingface_hub python package.

compatible means the Api should reuse the same files skipping downloads if they are already present and whenever this crate downloads or modifies this cache it should be consistent with huggingface_hub

At this time only a limited subset of the functionality is present, the goal is to add new features over time. We are currently treating this as an internel/external tool, meaning we will are currently modifying everything at will for out internal needs. This will eventually stabilize as it matures to accomodate most of our needs.

If you’re interested in using this, you’re welcome to do it but be warned about potential changing grounds.

If you want to contribute, you are more than welcome.

However allowing new features or creating new features might be denied by lack of maintainability time. We’re focusing on what we currently internally need. Hopefully that subset is already interesting to more users.

How to use

Add the dependency

cargo add hf-hub  # --features tokio

tokio feature will enable an async (and potentially faster) API.

Use the crate:

use hf_hub::api::sync::Api;

let api = Api::new().unwrap();

let repo = api.model("bert-base-uncased".to_string());
let _filename = repo.get("config.json").unwrap();

// filename  is now the local location within hf cache of the config.json file

Modules

api

The actual Api to interact with the hub.

Structs

Cache

A local struct used to fetch information from the cache folder.

CacheRepo

Shorthand for accessing things within a particular repo

Repo

The representation of a repo on the hub.

Enums

RepoType

The type of repo to interact with